python/requirements.txt
pandas
numpy
faker
google-cloud-bigquery
pyarrow
db-dtypes
google-auth